requirements/common.txt
Flask==0.12.2
flask-sqlalchemy
flask-migrate
flask-restful
flask-jwt
flask-testing
flask-mail
pymysql
